fix(webhook): restrict runner provider to ec2
